update to compile ocudu_mac lib
	modified:   CMakeLists.txt
	modified:   include/ocudu/adt/unique_function.h
	modified:   include/ocudu/asn1/asn1_utils.h
	modified:   include/ocudu/ran/ssb/ssb_properties.h
	modified:   include/ocudu/support/cpu_architecture_info.h
	modified:   include/ocudu/support/memory_pool/bounded_object_pool.h
	modified:   include/ocudu/support/resource_usage/resource_usage_utils.h
	modified:   lib/CMakeLists.txt
	modified:   lib/asn1/asn1_utils.cpp
	modified:   lib/mac/mac_dl/mac_subframe_time_mapper_impl.cpp
	modified:   lib/mac/mac_sched/ocudu_scheduler_adapter.cpp
	modified:   lib/mac/mac_ul/mac_ul_sch_pdu.h
	modified:   lib/ocudulog/event_trace.cpp
	modified:   lib/ocudulog/formatters/text_formatter.cpp
	modified:   lib/ocudulog/sinks/udp_sink.h
	modified:   lib/pcap/backend_pcap_writer.cpp
	modified:   lib/support/cpu_architecture_info.cpp
	modified:   lib/support/executors/unique_thread.cpp
	modified:   lib/support/network/io_broker_epoll.cpp
	modified:   lib/support/network/io_timer_source.cpp
	modified:   lib/support/network/sockets.cpp
	modified:   lib/support/resource_usage/perf_event_powercap_reader_impl.cpp
	modified:   lib/support/resource_usage/perf_event_powercap_reader_impl.h
	modified:   lib/support/resource_usage/resource_usage_utils.cpp
	modified:   lib/support/synchronization/futex_util.cpp
	modified:   lib/support/tracing/event_tracing.cpp
	modified:   lib/support/tracing/resource_usage.cpp

+++ lib/ocudulog/event_trace.cpp
@@ -86,7 +86,7 @@ void trace_duration_begin(const std::string& category, const std::string& name)
 
   char fmt_time[24];
   format_time(fmt_time, sizeof(fmt_time));
-  (*tracer)("[{}] [TID:{}] Entering \"{}\": {}", fmt_time, (unsigned)::pthread_self(), category, name);
+  (*tracer)("[{}] [TID:{}] Entering \"{}\": {}", fmt_time, (uintptr_t)::pthread_self(), category, name);
 }
 
 void trace_duration_end(const std::string& category, const std::string& name)
@@ -97,7 +97,7 @@ void trace_duration_end(const std::string& category, const std::string& name)
 
   char fmt_time[24];
   format_time(fmt_time, sizeof(fmt_time));
-  (*tracer)("[{}] [TID:{}] Leaving \"{}\": {}", fmt_time, (unsigned)::pthread_self(), category, name);
+  (*tracer)("[{}] [TID:{}] Leaving \"{}\": {}", fmt_time, (uintptr_t)::pthread_self(), category, name);
 }
 
 } // namespace ocudulog
